Added fy driver.
This works with FY2300, FY6600, FY6800, FY6900 and probably more

The main difference is that it uses read/retries instead of a fixed
timeout.  This speeds up commands.  In measurement, a 50 point bode
plot sped up from 112 seconds to 68 seconds.

I named the driver fy.py to keep the fy6600.py file intact and to
make it clearer that the driver is not limited to the fy6600.  We
could also set things up differently, if you would prefer.